update/wallet type change to decimal

+++ boredBets/boredBets/Models/Dtos/UserDto.cs
@@ -2,8 +2,8 @@
 {
     public record UserCreateDto(string Email,string Username, string Password);
     public record UserLoginDto(string EmailOrUsername, string Password);
-    public record UserDto(Guid Id,string Email, string Password,string Username, int Wallet, bool? Admin);
-    public record UserWalletDto(int Wallet);
+    public record UserDto(Guid Id,string Email, string Password,string Username, decimal Wallet, bool? Admin);
+    public record UserWalletDto(decimal Wallet);
 
     public record UsernameDto(string Username);
     public record UserEmailDto(string Email);
